Project UAS/requirements.txt
opencv-python
streamlit
Matplotlib
numpy
pillow